Editted all the classes
The GuessWord class was updated to ask the user for name and creates a player object. The Dict class was edited to remove the unwanted print statements. A print method was added to the player class. The Game class was edited to print end message if user wins or not